Reworked handler generation for super commands
Changed the approach of ArrayLists to simply calling the low level
handlers in order. The assumption is that it works and it will call each
handler in sequence and wait for the first one to be done before calling
the next one and so on, until it is done. Untested, but hopefully it
works. Will require testing on old project to see if it works.